flac: Disable VSX completely on ppc64
TODO Maybe make this configurable? Untested post CMake switch.

Upstream has a check for `vec_doubleh` and enables the use of VSX instructions
if present in altivec.h. This works fine with a GCC that doesn't try to optimise
for an specific CPU, but breaks on CPUs that
- have AltiVec instructions and
- do not have VSX instructions.
i.e. the G5. So let's just explicitly disable VSX for now when we're on 64-bit
big-endian POWER.