address-book-service: Fix galera plugin's linkage to libphonenumber
The galera engine plugin pulls in objects that expect libphonenumber symbols, but the plugin isn't getting linked against it.
When qtpim tries to load the plugin, it fails due to the missing symbol.
It subsequently does insufficient checks against such a loading error, and dereferences a NULL pointer when it iterates over previously found engines.