x265: Fix SIGILL on non-POWER8 ppc64
The upstream CMake option `CPU_POWER8` forces `-mcpu=power8` CPU target and enables handling of `ENABLE_ALTIVEC`.
But it defaults to on if it detects any POWER target - not specifically a POWER8 or higher family chip, or a target that
guarantees support for POWER8 instructions.
Only ppc64le guarantees POWER8+, ppc64 can be much older. The -mcpu= compiler flag causes SIGILL errors when trying to use
an x265 compiled with `CPU_POWER8` on an older CPU, so only enable the flag on ppc64le and disable it otherwise.

Additionally `ENABLE_ALTIVEC` is also default-enabled on POWER and tries to use VSX intrinsics, which are supported on POWER7+.
It already gets disabled for the high bit depth builds, but is left on its default value during the main build.
So only enable this option in the main build on ppc64le as well, and disable it otherwise.