{spidermonkey_115,spidermonkey_128,spidermonkey_140}: Pass stdenv.{build,host}Platform.rust.rustcTarget
Spidermonkey tries to figure out an appropriate triplet for Rust based on the C compiler host & target triplets we give it.
This fails when our triplets cannot be understood as-is by LLVM (gnuabielfv{1,2} on POWER).

We already have a mechanism to convert our triplets for Rust usage, so tell Spidermonkey to use the converted triplets.